+++ scripts/build-plugins.sh
@@ -53,7 +53,7 @@ for stack in "${STACKS[@]}"; do
 done
 
 # Standalone plugins (not language stacks): each builds plugins/src/<name> -> plugins/lisa-<name>
-STANDALONE=(wiki)
+STANDALONE=(wiki openclaw)
 for name in "${STANDALONE[@]}"; do
   build_plugin "$name" "lisa-$name"
 done